+++ chromadb/test/ef/test_onnx_mini_lm_l6_v2.py
@@ -8,8 +8,11 @@ import pytest
 import onnxruntime
 from unittest.mock import patch, MagicMock
 
-from chromadb.embedding_functions import ONNXMiniLM_L6_V2
-from chromadb.embedding_functions.embedding_function import Space, EmbeddingFunction
+from chromadb.utils.embedding_functions import ONNXMiniLM_L6_V2
+from chromadb.utils.embedding_functions.embedding_function import (
+    Space,
+    EmbeddingFunction,
+)
 
 
 class TestONNXMiniLM_L6_V2:
