+++ chromadb/utils/embedding_functions/roboflow_embedding_function.py
@@ -1,9 +1,150 @@
-# This file is maintained for backward compatibility
-# It imports and re-exports the RoboflowEmbeddingFunction from the new location
-
-from chromadb.embedding_functions.roboflow_embedding_function import (
-    RoboflowEmbeddingFunction,
+from chromadb.utils.embedding_functions.embedding_function import (
+    EmbeddingFunction,
+    Space,
 )
+from chromadb.api.types import (
+    Documents,
+    Embeddings,
+    Images,
+    is_document,
+    is_image,
+    Embeddable,
+)
+from typing import List, Dict, Any, Union, cast, Optional
+import os
+import importlib
+import base64
+from io import BytesIO
+import numpy as np
 
-# Re-export everything for backward compatibility
-__all__ = ["RoboflowEmbeddingFunction"]
+
+class RoboflowEmbeddingFunction(EmbeddingFunction[Embeddable]):
+    """
+    This class is used to generate embeddings for a list of texts or images using the Roboflow API.
+    """
+
+    def __init__(
+        self,
+        api_key_env_var: str = "ROBOFLOW_API_KEY",
+        api_url: str = "https://infer.roboflow.com",
+        api_key: Optional[str] = None,
+    ) -> None:
+        """
+        Create a RoboflowEmbeddingFunction.
+
+        Args:
+            api_key_env_var (str, optional): Environment variable name that contains your API key for the Roboflow API.
+                Defaults to "ROBOFLOW_API_KEY".
+            api_url (str, optional): The URL of the Roboflow API.
+                Defaults to "https://infer.roboflow.com".
+        """
+        # try:
+        #     import httpx  # type: ignore
+        # except ImportError:
+        #     raise ValueError(
+        #         "The httpx python package is not installed. Please install it with `pip install httpx`"
+        #     )
+
+        self.api_key_env_var = api_key_env_var
+        self.api_key = api_key or os.getenv(api_key_env_var)
+        if not self.api_key:
+            raise ValueError(f"The {api_key_env_var} environment variable is not set.")
+
+        self.api_url = api_url
+
+        try:
+            self._PILImage = importlib.import_module("PIL.Image")
+        except ImportError:
+            raise ValueError(
+                "The PIL python package is not installed. Please install it with `pip install pillow`"
+            )
+
+        self._httpx = importlib.import_module("httpx")
+
+    def __call__(self, input: Embeddable) -> Embeddings:
+        """
+        Generate embeddings for the given documents or images.
+
+        Args:
+            input: Documents or images to generate embeddings for.
+
+        Returns:
+            Embeddings for the documents or images.
+        """
+        embeddings = []
+
+        for item in input:
+            if is_image(item):
+                image = self._PILImage.fromarray(item)
+
+                buffer = BytesIO()
+                image.save(buffer, format="JPEG")
+                base64_image = base64.b64encode(buffer.getvalue()).decode("utf-8")
+
+                infer_clip_payload_image = {
+                    "image": {
+                        "type": "base64",
+                        "value": base64_image,
+                    },
+                }
+
+                res = self._httpx.post(
+                    f"{self.api_url}/clip/embed_image?api_key={self.api_key}",
+                    json=infer_clip_payload_image,
+                )
+
+                result = res.json()["embeddings"]
+                embeddings.append(np.array(result[0], dtype=np.float32))
+
+            elif is_document(item):
+                infer_clip_payload_text = {
+                    "text": item,
+                }
+
+                res = self._httpx.post(
+                    f"{self.api_url}/clip/embed_text?api_key={self.api_key}",
+                    json=infer_clip_payload_text,
+                )
+
+                result = res.json()["embeddings"]
+                embeddings.append(np.array(result[0], dtype=np.float32))
+
+        # Cast to the expected Embeddings type
+        return cast(Embeddings, embeddings)
+
+    @staticmethod
+    def name() -> str:
+        return "roboflow"
+
+    def default_space(self) -> Space:
+        return Space.COSINE
+
+    def supported_spaces(self) -> List[Space]:
+        return [Space.COSINE, Space.L2, Space.INNER_PRODUCT]
+
+    @staticmethod
+    def build_from_config(
+        config: Dict[str, Any]
+    ) -> "EmbeddingFunction[Union[Documents, Images]]":
+        api_key_env_var = config.get("api_key_env_var")
+        api_url = config.get("api_url")
+
+        if api_key_env_var is None or api_url is None:
+            assert False, "This code should not be reached"
+
+        return RoboflowEmbeddingFunction(
+            api_key_env_var=api_key_env_var, api_url=api_url
+        )
+
+    def get_config(self) -> Dict[str, Any]:
+        return {"api_key_env_var": self.api_key_env_var, "api_url": self.api_url}
+
+    def validate_config_update(
+        self, old_config: Dict[str, Any], new_config: Dict[str, Any]
+    ) -> None:
+        # API URL can be changed, so no validation needed
+        pass
+
+    def validate_config(self, config: Dict[str, Any]) -> None:
+        # TODO: Validate with JSON schema
+        pass
